Reorganize allocators and buffers into modules
