--- nv-video-codec/src/decoder/decoding_output.rs
@@ -1,7 +1,0 @@
-use crate::decoder::FrameInfo;
-
-pub struct DecodingOutput<I> {
-    pub frames: I,
-    pub frame_count: usize,
-    pub frame_info: FrameInfo,
-}
