+++ nv-video-codec/src/decoder/mod.rs
@@ -1,18 +1,13 @@
 pub mod builder;
-pub mod decoding_output;
 pub mod error;
 pub mod flags;
 pub mod frame;
-pub mod frame_info;
 pub mod nvdecoder;
 pub mod types;
 pub mod videoformat;
 
 pub use builder::*;
-pub use decoding_output::*;
 pub use error::*;
 pub use flags::*;
-pub use frame::*;
-pub use frame_info::*;
 pub use nvdecoder::*;
 pub use videoformat::*;
